requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
atlassian-python-api==3.41.13
attrs==23.2.0
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
dataclasses-json==0.6.6
Deprecated==1.2.14
frozenlist==1.4.1
idna==3.7
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==2.4
langchain==0.2.1
langchain-community==0.2.1
langchain-core==0.2.1
langchain-text-splitters==0.2.0
langsmith==0.1.63
marshmallow==3.21.2
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.26.4
oauthlib==3.2.2
orjson==3.10.3
packaging==23.2
pydantic==2.7.1
pydantic_core==2.18.2
python-dotenv==1.0.1
PyYAML==6.0.1
requests==2.32.2
requests-oauthlib==2.0.0
six==1.16.0
soupsieve==2.5
SQLAlchemy==2.0.30
tenacity==8.3.0
typing-inspect==0.9.0
typing_extensions==4.12.0
urllib3==2.2.1
wrapt==1.16.0
yarl==1.9.4
